player operations -> player panel
